requirements.txt
numpy
pandas
scikit-learn==1.5.1
streamlit
Pillow
joblib
seaborn
matplotlib
watchdog
